+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/supplies/SupplyRepository.kt
@@ -1,10 +1,13 @@
 package com.app.arcabyolimpo.domain.repository.supplies
 
+import android.net.Uri
 import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
 import com.app.arcabyolimpo.domain.model.filter.FilterData
 import com.app.arcabyolimpo.domain.model.supplies.Batch
 import com.app.arcabyolimpo.domain.model.supplies.Supply
+import com.app.arcabyolimpo.domain.model.supplies.SupplyAdd
 import com.app.arcabyolimpo.domain.model.supplies.SupplyBatchExt
+import com.app.arcabyolimpo.domain.model.supplies.WorkshopCategoryList
 
 /**
  * Retrieves detailed information for a specific supply by its [id].
@@ -28,4 +31,10 @@ interface SupplyRepository {
 
     // Yo uso Update
     suspend fun deleteOneSupply(id: String)
+    suspend fun getWorkshopCategoryList(): Result<WorkshopCategoryList>
+
+    suspend fun addSupply(
+        supply: SupplyAdd,
+        image: Uri?
+    ): Result<Unit>
 }