+++ app/src/main/java/com/app/arcabyolimpo/domain/usecase/ExternalCollaborator/ExternalCollabsUseCase.kt
@@ -4,10 +4,10 @@ import com.app.arcabyolimpo.domain.model.ExternalCollaborator.ExternalCollab
 import com.app.arcabyolimpo.domain.repository.ExternalCollabRepository.ExternalCollabRepository
 import javax.inject.Inject
 
-class GetAllCollabsUseCase @Inject constructor(
-    private val repository: ExternalCollabRepository
-) {
-    suspend operator fun invoke(): Result<List<ExternalCollab>> {
-        return repository.getAllCollabs()
+class GetAllCollabsUseCase
+    @Inject
+    constructor(
+        private val repository: ExternalCollabRepository,
+    ) {
+        suspend operator fun invoke(): Result<List<ExternalCollab>> = repository.getAllCollabs()
     }
-}